Refactored Main App Layout
