+++ src/javascript/app/components/Gallery/index.tsx
@@ -6,7 +6,8 @@ import GalleryGrid from '../GalleryGrid';
 import GalleryGroup from '../GalleryGroup';
 import GalleryHeader from '../GalleryHeader';
 import GalleryImage from '../GalleryImage';
-import GalleryIntro from '../GalleryIntro';
+import GalleryNumbers from '../GalleryNumbers';
+import StorageWarning from '../StorageWarning';
 import Pagination from '../Pagination';
 import { useGallery } from './useGallery';
 
@@ -25,9 +26,10 @@ function Gallery() {
   return (
     <Stack
       direction="column"
-      gap={2}
+      gap={1}
     >
-      <GalleryIntro
+      <StorageWarning />
+      <GalleryNumbers
         imageCount={imageCount}
         selectedCount={selectedCount}
         filteredCount={filteredCount}
@@ -35,32 +37,37 @@ function Gallery() {
       { enableImageGroups ? (
         <FolderBreadcrumb />
       ) : null }
-      <GalleryHeader page={page} isSticky />
-      <Pagination page={page} />
-      <GalleryGrid>
-        { images.map((image) => (
-          covers.includes(image.hash) ? (
-            <GalleryGroup
-              key={image.hash}
-              hash={image.hash}
-            />
-          ) : (
-            <GalleryImage
-              key={image.hash}
-              hash={image.hash}
-              page={page}
-            />
+      <Stack
+        direction="column"
+        gap={2}
+      >
+        <GalleryHeader page={page} isSticky />
+        <Pagination page={page} />
+        <GalleryGrid>
+          { images.map((image) => (
+            covers.includes(image.hash) ? (
+              <GalleryGroup
+                key={image.hash}
+                hash={image.hash}
+              />
+            ) : (
+              <GalleryImage
+                key={image.hash}
+                hash={image.hash}
+                page={page}
+              />
+            )
+          )) }
+        </GalleryGrid>
+        {
+          images.length < 3 ? null : (
+            <>
+              <Pagination page={page} />
+              <GalleryHeader page={page} isBottom />
+            </>
           )
-        )) }
-      </GalleryGrid>
-      {
-        images.length < 3 ? null : (
-          <>
-            <Pagination page={page} />
-            <GalleryHeader page={page} isBottom />
-          </>
-        )
-      }
+        }
+      </Stack>
     </Stack>
   );
 }
