--- src/javascript/app/components/Layout/index.scss
@@ -1,58 +1,0 @@
-.layout {
-  width: $layout-width-6-cols; // Size is based on padding/gap/imagesize of gallery grid with 6 images scaled 1x
-  margin: 0 auto;
-  padding: 120px var(--layout-padding-side) 20px;
-  position: relative;
-  z-index: 0;
-  min-height: 100dvh;
-
-  --1x-columns: 6;
-  --2x-columns: 3;
-  --max-columns: 1;
-
-  --1x-gap: 16px;
-  --2x-gap: 32px;
-  --max-gap: 32px;
-
-  --layout-padding-side: 20px;
-
-  @media (max-width: $breakpoint-5-cols) {
-    width: $layout-width-5-cols;
-    --1x-columns: 5;
-    --2x-columns: 2;
-  }
-
-  @media (max-width: $breakpoint-4-cols) {
-    width: $layout-width-4-cols;
-    --1x-columns: 4;
-    --2x-columns: 2;
-  }
-
-  @media (max-width: $breakpoint-3-cols) {
-    width: $layout-width-3-cols;
-    --1x-columns: 3;
-    --2x-columns: 1;
-    --layout-padding-side: 10px;
-  }
-
-  @media (max-width: $breakpoint-2-cols) {
-    width: $layout-width-2-cols;
-    --1x-columns: 2;
-    --layout-padding-side: 5px;
-  }
-
-  @media (max-width: $breakpoint-1-cols) {
-    width: $layout-width-1-cols;
-    --1x-columns: 1;
-  }
-
-  &__main-headline {
-    font-size: 30px;
-    margin-bottom: 30px;
-    @include theme-color($color-primary, $color-primary-bright);
-
-    &:not(:first-child) {
-      margin-top: 60px;
-    }
-  }
-}
